# Speed up no-op installs