@segiddins 2017-07-11
This is ready, please review @indirect.